[axl_sys_win] fix: RegCloseKey returns LONG (error), not BOOL (success)

+++ src/axl_io/axl_io_win_SerialPortEnumerator.cpp
@@ -74,7 +74,7 @@ SerialPortEnumerator::createPortList(sl::List<SerialPortDesc>* portList)
 SerialPortDesc*
 SerialPortEnumerator::createPortDesc(sys::win::DeviceInfo* deviceInfo)
 {
-	int result;
+	long result;
 
 	sys::win::RegKeyHandle devRegKey = deviceInfo->openDeviceRegistryKey(KEY_QUERY_VALUE);
 	if (devRegKey == INVALID_HANDLE_VALUE)
